% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_requests_count.R
\name{get_requests_count}
\alias{get_requests_count}
\title{Retrieve your account's remaining request limit - for the day}
\usage{
get_requests_count()
}
\value{
A list, representing the response from the tiny.cc API.
}
\description{
tbd.
}
\details{
tbd.
}
\examples{

}
\seealso{
\code{<https://tiny.cc/api-docs>}
}
\author{
Philipp Ottolinger
}
